refactor auth logic into hooks
